% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldcupfunc.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Generate Population Trends Graph for a Specific Country}
\usage{
CountryPopulation(country_name, title = NULL)
}
\arguments{
\item{country_name}{A character string specifying the country for which the
population trends graph should be generated.}

\item{title}{An optional parameter specifying the title of the graph. If not
provided, a default title will be used.}
}
\value{
A ggplot object representing the population trends graph for the
        specified country.
}
\description{
This function takes a country name and generates a population trends graph
over the years from 1950 to 2020 using the WorldPopulation data.
}
\details{
The function filters the WorldPopulation data for the specified country,
pivots the data to long format, and creates a ggplot object with a point plot
showing the population trends over the years.
}
\examples{
# Example usage of the dataset

head(WorldPopulation)

CountryPopulation("Ghana")

# Additional examples go here (WorldPopulation Site)
}
