% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldcupdoc.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Scrape and Clean FIFA World Cup Attendance Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 5 columns.
}
\usage{
World_Cup
}
\value{
A data frame containing FIFA World Cup attendance data.
}
\description{
This function reads the HTML content of the FIFA World Cup Wikipedia page,
extracts the attendance data from a specific table, cleans and formats the
data, and then stores it in a data frame.
}
\details{
The function uses the rvest package to scrape the FIFA World Cup Wikipedia
page and extract attendance data from a specific table. The data is then
cleaned and formatted, and unnecessary rows are removed. The resulting data
frame is stored in the `Attendance` object.


Run this example only in interactive mode
url <- read_html("https://en.wikipedia.org/wiki/FIFA_World_Cup")




 @examples
 Example usage of the dataset

head(World_Cup)

 Additional information visit wikipedia website
}
\keyword{datasets}
