% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpopdoc.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Data Set}
\format{
A data frame with variables:
\describe{
  \item{Year}{The year of the population estimate.}
  \item{Country Name}{The name of the country.}
  \item{Population}{The population estimate for the corresponding year and country.}
}
}
\source{
United Nations World Population Prospects.
}
\usage{
WorldPopulation
}
\description{
This data set includes population estimates for countries from 1950 to 2020.
}
\examples{
# Example usage of the dataset

head(WorldPopulation)

# Additional information visit World Population website
}
\seealso{
\code{\link{YaokumahWorldPopulation::WorldPopulation}}
}
\keyword{datasets}
